SEQUOIADBMAINSTREAM-3199:【BUG】【new】【配置动态生效】auditmask、startshifttime不能动态生效，hjbuf设值过大内置sql查询时会导致节点重启 修复hjbuf的bug
git-svn-id: http://192.168.20.11/sequoiadb/branches/sequoiadb_2.6@34065 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/qgm/qgmPlHashJoin.cpp
@@ -138,7 +138,7 @@ namespace engine
 
       _hitBuildEnd = FALSE ;
 
-      UINT64 bufSize = pmdGetOptionCB()->getHjBufSize() * 1024 * 1024 ;
+      UINT64 bufSize = ( ( UINT64 )pmdGetOptionCB()->getHjBufSize() ) * 1024 * 1024 ;
       rc = _hashTbl.init( bufSize ) ;
       if ( SDB_OK != rc )
       {
